---------------------------- in mj514.c/.h: - worked around the EXTI3 problem when the motor controller is enabled - added PA5 - 12V0 regulator shutdown pin (preparation for HW rev. 1a) - added one more debug GPIO - sorted out the interrupt priority mess
in i2c.c/.h:
- added switch for DMA vs. IT calls
- added a define which controls HAL_I2C_IsDeviceReady()

in motor.c/.h:
- added basic angle & current statistics
- added #if USE_AS5601_PULSE to deactivate rotary encoder A/B channel code
- added #define USE_RAMPUP so that rampup code can be disabled
- in timer16 ISR: added code which allows timer16 to continue to run after end travel is reached
- in _StartMotorController() & _StopMotorController() added timer16 interrupt priority change (suspect was some race condition or priority collision)
- added typecast to get rid of compiler warnings
- fixed motor fault EXTI handling

in mj514_adc.c:
- added typecasts to avoid compiler warnings

in gear.c:
- added a while() loop to have some from of atomic shifting

in mb85rc.c:
- modified code so that this driver can work with 2-byte wide FeRAM register addresses

in button.c & composite_led_actual.c:
- added #if defined(USE_BUTTON) and USE_LED

in event.c:
- replaced e.g. #if (MJ808_) with #if defined(MJ808_)
- added typecast to get rid of compiler warnings

in try.c:
- replaced e.g. #if (MJ808_) with #if defined(MJ808_)
- added typecast to get rid of compiler warnings

added an xls with calculations of precise decimal pulses/raw angles and their next rounded number, along with corrections

+++ mj8x8/Core/Inc/mb85rc/mb85rc.h
@@ -7,8 +7,8 @@
 
 typedef struct	// struct describing the FeRAM functionality
 {
-	uint32_t (*Read)(const uint8_t RegAddr, const uint8_t size);  // reads one byte of data out of the FeRAM module from given address
-	void (*Write)(const uint32_t data, const uint8_t RegAddr, const uint8_t size);  // writes one byte of data into the FeRAM module at given address
+	uint32_t (*Read)(const uint16_t RegAddr, const uint8_t size);  // reads one byte of data out of the FeRAM module from given address
+	void (*Write)(const uint32_t data, const uint16_t RegAddr, const uint8_t size);  // writes one byte of data into the FeRAM module at given address
 } mb85rc_t;
 
 mb85rc_t* mb85rc_ctor(void);	// the MB85RC constructor
